refactor : form 요소 뷰 - 로직 분리했습니다
